Only copy to latest if it's actually the latest

+++ .github/handle_addon_file.py
@@ -3,6 +3,8 @@ import sys
 
 # versions dir
 versions_dir = pathlib.Path("versions")
+
+
 latest_dir = versions_dir / "latest"
 # create the versions directory if it doesn't exist
 if not versions_dir.exists():
@@ -19,7 +21,7 @@ if addon_file is None:
     print("No .mcaddon file found in the main directory")
     sys.exit(1)
 
-#print(f"Found {addon_file.name}, {addon_file}")
+# print(f"Found {addon_file.name}, {addon_file}")
 *names, version = addon_file.stem.split(" ")
 if not names:
     *names, version = addon_file.stem.split("_")
@@ -28,10 +30,10 @@ if not names or not version:
     print(f"Invalid file name: {addon_file.name}")
     sys.exit(1)
 
-#print(
+# print(
 #    "names:", names,
 #    "version:", version
-#)
+# )
 major, minor, *patch = version.split(".")
 # create a dir in major.patch if the patch exists
 if patch:
@@ -47,7 +49,7 @@ else:
         version_dir.mkdir()
         #print(f"Created {version_dir}")
 
-#print(f"version: {version}")
+# print(f"version: {version}")
 
 # change the file ext to .zip
 new_file = addon_file.with_suffix(".zip")
@@ -62,19 +64,27 @@ with zipfile.ZipFile(new_file, "r") as zip_ref:
     zip_ref.extractall(version_dir)
 
 new_file.unlink()
-#print(f"Deleted {new_file}")
+# print(f"Deleted {new_file}")
 
+versions = versions_dir.glob("*")
+# remove the latest directory from the list
+versions = sorted([version for version in versions if version.name != "latest"])
 
-# copy the files to the latest directory as well
-import shutil
-shutil.rmtree(latest_dir, ignore_errors=True)
-shutil.copytree(new_dir, latest_dir)
+latest_version = versions[-1]
+previous_version = versions[-2]
+latest_version_number = float(latest_version.parts[-1])
+previous_version_number = float(previous_version.parts[-1])
 
-#print(f"Copied {new_dir} to {latest_dir}")
+if latest_version_number > previous_version_number:
+    import shutil
+    shutil.rmtree(latest_dir, ignore_errors=True)
+    shutil.copytree(new_dir, latest_dir)
+
+# print(f"Copied {new_dir} to {latest_dir}")
 # write version to txt file called version.txt
 github_dir = pathlib.Path(".github")
 
-#print(f"Done, handled {addon_file.name}")
+# print(f"Done, handled {addon_file.name}")
 
 if __name__ == "__main__":
     print(version)
